Add MySQL support to bendo server
Many other changes from adding this. Refactor QL (the embedded database)
to use the same interface so the application can switch between them,
using QL when MySQL is not configured.

This will put the item metadata cache into the database. It adds a
command line parameter to give information to configure mysql. It also
defines the new FixityDB interface and alters the fixity code to use
that instead. Fixity rate limiting is removed (for the moment) to help
simplify the code. The rate counter is moved into the util package for
the time being.

 * Up default cache size to 100 MB
 * Change file store to only list actual gettable files previously it
listed EVERY file inside its storage directory. now it only lists files
which are exactly two levels deep.
 * Change configuration of server to be more general.
Now the server will set up a database if needed, and it will also set up
the queue and cache directories given a single directory to put them in.

Still need to add REST interface to check on fixity status.

issue #10