Add ability to shut down the server gracefully
We use the httpdown package to provide a way to close the http socket
and exit when all the pending requests have finished. The critical
component in the shutdown, the transaction writting goroutines, are
handled by changing the gate code to track the goroutines inside it and
to release goroutines waiting to enter it. Finally, we add a signal
handler to the bendo server application.

* move transaction gate into the server struct
* move fixity scanner into the fixity file
* fix makefile rules to always build the targets
* add gate test code