Add transaction cleaner
It will run twice a day and remove any finished transactions.

* Change transaction serialization to fix transaction loading bug
(prevously the transaction state was saved as a string. but there was no
code to unserialize the string. Change to save the state as an integer)
* Delete successful transactions after 24 hours
* Delete unsuccessful transactions after 1 week
* Delete orphaned files after 2 weeks
* Change transaction html template to hyperlink uploaded files
* Remove tx: prefix for transaction stores. Since we are putting
transactions into their own subdirectory of the cache dir, prefixing the
keys with a "tx:" is not needed anymore.

Issue #39