Log which database we are using.
Also reformat the runit run script so the command line options are
clearer. The problem with bendo not connecting to mysql on the staging
machine was was the run script on the server was missing the
`--cache-size` option, so its parameter was interpreted as the end of the
options and then the `--mysql` option was not parsed.