Back out QL database migration
Doing the DROP COLUMN and the indexes was triggering a bug in the QL
library. Waiting until we have more time to look into this. It is also
not essential to the QL library since every table row is already
autoindexed via the id() function.

Also add sample config file and integration tests for mysql. They are
not perfect, but they are a start.