Change server transaction worker model
Previously each pending transaction was put on its own goroutine and
they were limited using a gate. This is hard to reason about. So this
commit changes it to have a pool of worker transaction goroutines and
they get the transactions to process from a channel.

The channel is buffered, so the source route won't block.
Also use a second cancel channel for the transaction workers to signal
when they should exit. (since closing the transation work channel won't
stop them until the channel is empty).

Add better error handling for missing files in the blob cache.