Handle errors getting cached content better
When there is an error getting cached content, pretend like the item is
not in the cache. This handles the case where our internal LRU list gets
out of sync with the files actually on the system. We handle it by
deleting any files present, so the next time we can cache the object.