Add mime-type to bclient
This was a fairly invasive change. The processing logic was split
between the cmd/bclient and the fileutil package, so I moved fileutil
into bclient. The processing was reworked to be more blob centric. We
now maintain a list of unique local blobs so if two files have the same
content, we only need to upload it once. We also had to add mime-type
sensitivity to the blob resolver since it is possible a blob is
unchanged except for the mime type, and then that will need updating.

The blob resolver now compares the local and remote trees and creates a
list of actions. This seemed more explicit and easier to adjust the
resolution algorithm. It also separates out the resolution algo for
testing (albeit, no tests are supplied for it).

Mime types are passed in by using files named `bclient-manifest` in the
file tree. These files also provide a way to pass precomputed md5 and
sha256 sums to the server, as part of the end-to-end hash checking.

bclientapi was adjusted slightly. Mostly by making chunksize implicit
(since it IS stored in the ItemAttributes structure).
It also needs to send the mime-type when a new blob is uploaded.

Most file handling was adjusted to use absolute paths, in an attempt to
reduce the number of calls to path.Join.