Add time-based cache
A time based cache has an eviction policy to remove items that have not
been accessed for a set length of time. In particular, a time-based
cache may grow arbitrarily large. (however, for cloud storage that is
not necessarily a problem).

Altered the Cache interface slightly. Added MaxSize().
Refactored the internal writer to work with both cache classes.