Adding S3 store adapter
It tries not to make any assumptions on object size, and all data
operations are stream based. There is an integration test. Not sure how
to adapt the pieces, such as the writeCloser or the reader to work with
unit tests--a way to abstract out the direct S3 calls is needed.
I hope a future commit will do that.