Improve transaction error handling
* change the bundle writing process to make sure the blob list is in a
consistent state in the case of copy errors. Always make an entry for a
blob if it was created in a bundle file. Also return read errors when
copying a file into a bundle.

* change transaction processing to stop when the verify step found
errors. (previously it just logged them). Also make verify more
sensitive to whether there was a checksum mismatch or a general error.
This changes the fragment Verify() to return a (bool, error) pair
instead of only a bool.

Addresses DLTP-1676