Adding additional logging
In trouble shooting a tape storage issue, I'm fumbling as to where the
problem might be. I'm currently searching down the request path for
a bendo item.

I found these two places in which an error could be getting swallowed.
I don't believe this is the case, but for completeness, I'm adding the
logic.

I'm relying on Don to assess if this is a good idea to merge.