Refactor blob caching path
Since the blob caching path is very important, restructure it so it
singleflights blob caching when many people want the same blob. Also be
clearer with what is happening inline in the requests and what is
happening in the background. Finally, add more logging to the procedure
to help understand the issues we seem to be having with S3.