Add version, blob, and slot database tables.
The server will need a better way to search for the files, so we need to
index all the item metadata in their own tables.
This will let us do more in the future, such as
 * find all things that changed since a given date
 * find files based on mime-type
 * find files based on size
 * find files based on names (e.g. list every rof file)
 * more efficient file downloads

This commit only adds the new tables and code to populate item into
them. The above integrations will happen in subsequent commits.