WIP
doesnt compile. trying to apply the centralize-item-handling patch.
i think the goal is to remove the slot resolution completely from the
item handler and put it all in the server.

there is also something to singleflight the requests for tape. need to
figure out how it interacts with the blob caching single-flights. and
whether it is even possible/diserable. maybe it can wait for later?