Add BlackPearl storage driver
It interacts directly with a BlackPearl device via API.

Since the BlackPearl has no way to upload a file without knowing the
size, we use a local temporary file to store uploads before sending the
whole thing to the device.

To be done next is hooking the black pearl into the configuration code
so bendo can use it.