Add ability to configure BlackPearl endpoint
This adds a "blackpearl://" and "blackpearls://" schemes to the path
used to give the `StoreDir`. This lets a URL be given, and when that
scheme is given a BlackPearl driver is created.

This moves a lot of the server configuration from the server code into
cmd/bendo where it makes more sense. There were lots of fields passed
into the server that were only used at startup. Now the initial
executiable is responsible for reading the config file and bootstrapping
the server pieces.