Have S3 and BlackPearl stores make pages disjoint
This makes each page loaded align its start to a multiple of
defaultPageSize. This makes each possible offset in a file be located on
a unique page, facilitating the cache reuse. (e.g. reading a zip file
involves many reads from the end of a file, each with earlier offsets.)