Completely download files from BlackPearl
Before we would page in data as it was requested. This seemed to work
for small things, but was WAY too slow for large files. Also we would
see occasional errors that an item was not ready.

This changes the download process to always save the complete item to a
temp file space, and then reads are directed to the temp file.

There is a lot of room for future improvement. For example, every time
an item is Open() we download it completely. This will make things like
fixity checking slow since for fixity checking we Open() every file
associated with an item twice.