Make the upload-to-S3 buffers bigger
The goal is to decrease the numnber of API calls we make to s3 to upload
very large items, (e.g. > 1 GB). To do this we move the first upload
threshold from 5 MB to 64 MB, and then instead of increasing it by 128KB
each time, we double it until we reach an upload threshold of 4 GB. This
is close to the maximum multipart part size of 5 GB.

We'll need to see whether sending such large chunks to S3 causes us to
encounter more network errors. If so, maybe the threshold should stop
increasing at 2 GB or even 1 GB.