Add range requests for content downloads
Add a Seek() method to an internal adapter. This lets use use
http.ServeContent() and support range requests when downloading file
contents. This will work for files downloaded from our cache, but not
those downloaded directly from tape. (Since we cache ALL files in
production, nothing is ever served directly from tape, so this is moot).

This is part of the project to have end-to-end range requests for curate
downloads